Upgraded reslot to vuetify4
