+++ app/assets/devices.json
