+++ app/components/QrDialog.vue
