+++ app/components/TerminalLoader.vue
