+++ app/components/screen/Device.vue
