+++ package.json
@@ -10,19 +10,17 @@
     "postinstall": "nuxt prepare"
   },
   "dependencies": {
-    "@pinia-plugin-persistedstate/nuxt": "^1.2.1",
     "@pinia/nuxt": "^0.11.3",
     "nuxt": "^4.3.1",
     "pinia": "^3.0.4",
-    "qrcode.vue": "^3.8.0",
-    "vue": "^3.5.28",
-    "vue-router": "^4.6.4"
+    "qrcode.vue": "^3.10.0",
+    "vite-plugin-vuetify": "^2.1.3",
+    "vue": "^3.5.40",
+    "vue-router": "^4.6.4",
+    "vuetify": "^4.1.6"
   },
- 
   "devDependencies": {
     "@mdi/font": "^7.4.47",
-    "sass-embedded": "^1.97.3",
-    "vite-plugin-vuetify": "^2.1.3",
-    "vuetify": "^3.11.8"
+    "sass-embedded": "^1.100.0"
   }
 }
